chore(deps): Document current MSRV in Cargo.toml
This is the current defacto MSRV not an upgrade because the last release
with locked deps already requires this, it just wasn't specifically
documented. Documenting it should help in the future analizing what
release tags should be and whether or not to dump dependencies.

The current limitation comes via the `time` dependency burried in the
dependency tree.