chore!: Re-license under GNU Lesser General Public License v3
BREAKING CHANGE: As of this release the project is relicensed from GPL
to LGPL. From a consumer perspective this is a more permissive license
that will allow linking and building the various library iterations
without the viral nature of the GPL "infecting" the whole project.

All previous contributors have been contacted and approved of this
change. All future contributions will be accepted as LGPL.