refator!: Rename `to_*()` functions as just `*()`
BREAKING CHANGE: This renames the function in the Rust crate to match
the Lua/Python/JS bindings. It better follows some other crates that
have casing functions and traits that have a `to_` naming scheme. Since
we plan to implement a trait ourselves, that will differenciate the
trate functions from the global ones.