refactor(crate)!: Implement TryFrom instead of From to avoid panics
BREAKING CHANGE: The infalliable conversion functions are now mostly
falliable. This allows them to be used more robustly when the inputs
might not be know (e.g. the language identifier string hasn't been
validated or input has invalid UTF-8 encoding). The catch is that it
makes most function return a `Result<String>` instead of `String`.